HU08 Add ReportUtils
